Move also local docker build to this repository in order to remove it fro rails app

+++ docker/entrypoint.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-cd /app || exit
+cd /app
 
 if ! bundle exec rails runner "ActiveRecord::Base.connection rescue exit 1"
 then
@@ -8,10 +8,14 @@ then
     bundle exec rails db:create >log/latest-startup.log 2>&1
 fi
 
-bundle exec rails db:migrate >>log/latest-startup.log 2>&1
-bundle exec rails thecore:db:seed >>log/latest-startup.log 2>&1
-
-rm -f tmp/pids/server.pid
-bundle exec rails s -p 3000 -b '0.0.0.0'
+if bundle exec rails db:migrate >>log/latest-startup.log 2>&1
+then 
+    if bundle exec rails thecore:db:seed >>log/latest-startup.log 2>&1
+    then
+        # Only if all the migrations are ok, run the server
+        rm -f tmp/pids/server.pid
+        bundle exec rails s -p 3000 -b '0.0.0.0'
+    fi
+fi
 
 exit 0
